requirements.txt
langchain>=0.3.20
langchain-google-genai>=2.0.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
rdflib>=7.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pandas>=1.5.0
scipy>=1.10.0
scikit-learn>=1.3.0
numpy>=1.24.0
